pyproject.toml
[tool.poetry]
name = "pyedgarai"
version = "0.8.0"
description = "A package for retrieving SEC filings and parse them using LLMs"
authors = ["Juan F. Imbet"]
license = "MIT"
readme = "README.md"
packages = [{include = "pyedgarai", from = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
tqdm = "^4.66.5"
bs4 = "^0.0.2"
yfinance = "^0.2.43"
openpyxl = "^3.1.5"
flask-openapi3 = "^4.0.1"
openai = "^1.51.2"
sec-cik-mapper = "^2.1.0"
markdownify = "^0.14.1"
outlines = "^0.1.11"
transformers = "^4.47.1"
datasets = "^3.2.0"
scikit-learn = "^1.6.0"
pydantic = "^2.8.0"
requests = "^2.32.3"
pandas = "^2.2.2"

[tool.poetry.dev-dependencies]

[tool.poetry.group.dev.dependencies]
pytest-cov = "^5.0.0"
python-semantic-release = "^9.8.6"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.semantic_release]
version_variable = "pyproject.toml:version"
version_toml = [
    "pyproject.toml:tool.poetry.version"
]
build_command = "echo version = '%NEW_VERSION%' > version.py"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "--ignore=archive/"

